WifiStrength class only works if don't import WifiSignal class?
